+++ demo/demo_logger.py
@@ -1,8 +1,8 @@
 # 2017.8.14 small program to test fish logger.py
 # 2018.2.3 small edit, change file_file to file
 
-from fish_base.fish_logger import *
-from fish_base.fish_file import *
+from fishbase.fish_logger import *
+from fishbase.fish_file import *
 
 log_abs_filename = get_abs_filename_with_sub_path('log', 'fish_test.log')[1]
