+++ docs/change_log.rst
@@ -1,6 +1,10 @@
 更新记录
 ===========================
 
+2018.5.21 v1.0.12
+
+* 19035, edit fish_base to fishbase
+
 2018.5.18 v1.0.11
 ---------------------------
