+++ docs/index.rst
@@ -3,12 +3,12 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-.. image:: https://api.travis-ci.org/chinapnr/fish_base.svg?branch=v1.0.11_develop
+.. image:: https://api.travis-ci.org/chinapnr/fishbase.svg?branch=master
 
-fish_base 函数库使用说明
+fishbase 函数库使用说明
 =====================================
 
-**fish_base** 是我们自己开发、收集和整理的一套 Python 基础函数库。
+**fishbase** 是我们自己开发、收集和整理的一套 Python 基础函数库。
 抽象了很多常见的通用的业务逻辑，以希望减少日常开发中的一些重复的工作量。
 
 .. toctree::
@@ -30,12 +30,6 @@ API 函数列表
    fish_csv
    fish_logger
 
-----
 
-从 ``v1.0.11`` 开始，所有 fish_base 包的的引用需要用如下格式::
-
-    from fish_base.fish_aaaa import *
-
-也就是所有的 fish_base 子模块也是 ``fish_`` 开头。
