--- fish_base.egg-info/PKG-INFO
@@ -1,17 +1,0 @@
-Metadata-Version: 1.1
-Name: fish-base
-Version: 1.0.11
-Summary: some useful functions for python
-Home-page: https://github.com/chinapnr/fish_base
-Author: David Yi
-Author-email: wingfish@gmail.com
-License: MIT
-Description: UNKNOWN
-Platform: UNKNOWN
-Classifier: Development Status :: 3 - Alpha
-Classifier: Intended Audience :: Developers
-Classifier: Topic :: Software Development :: Build Tools
-Classifier: License :: OSI Approved :: MIT License
-Classifier: Programming Language :: Python :: 2.7
-Classifier: Programming Language :: Python :: 3.5
-Classifier: Programming Language :: Python :: 3.6
