--- fish_base.egg-info/SOURCES.txt
@@ -1,28 +1,0 @@
-LICENSE.txt
-MANIFEST.in
-README.md
-setup.py
-demo/.DS_Store
-demo/demo_common.py
-demo/demo_csv.py
-demo/demo_date.py
-demo/demo_file.py
-demo/demo_logger.py
-demo/test_conf.ini
-demo/csv/test_csv.csv
-demo/log/blank.txt
-demo/log/fish_test.log
-demo/test_conf/test_conf.ini
-dev_doc/fish_issues.md
-fish_base/__init__.py
-fish_base/fish_common.py
-fish_base/fish_csv.py
-fish_base/fish_date.py
-fish_base/fish_file.py
-fish_base/fish_logger.py
-fish_base.egg-info/PKG-INFO
-fish_base.egg-info/SOURCES.txt
-fish_base.egg-info/dependency_links.txt
-fish_base.egg-info/requires.txt
-fish_base.egg-info/top_level.txt
-test/test_common.py
