--- fish_base.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
