+++ test/test_common.py
@@ -5,7 +5,7 @@
 import pytest
 import sys
 sys.path.append('../fish_base')
-from fish_base.fish_common import *
+from fishbase.fish_common import *
 
 
 # 2018.5.14 v1.0.11 #19027 create by David Yi, 开始进行单元测试
